Backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
xgboost
psycopg2-binary
joblib
kafka-python
requests
